C++: Introduce an SSA pruning stage. In the next commit we will introduce the "real" SSA analysis that we'll end up using in dataflow (which handles) indirections. However, this will introduce many additional reads and writes. So to avoid this blowing up we will first perform a SSA stage that conflates &x, x, *x, etc.
The final relation we will use from this stage is the
'Def' class defined in the very bottom. This class has
been pruned to only contain those definitions which are
eventually read.

Both SSA stages uses a common 'SsaInternalsCommon' file
that specifies which indirection of a variable a given
store writes to, and which indirection a given read
reads from. However, this number isn't used in this
first SSA stage. It will, however, be used in the next
stage.