C++: Update field flow. This is mostly straightforward, but there are some subtleties. In addition to keeping track of indices on SSA read/writes, and dataflow nodes, we also need to keep indices on field contents. This is because we don't want dataflow in this example: ```cpp struct S { int* x; }; ... S s; *s.x = source(); sink(s.x) // reads the pointer, but not the pointee. ``` Additionally, now we've done all these smart things for field updates on non-union objects, but none of it works for union fields since these were previously excluded by our field flow. The only thing needed for proper dataflow through unions were a read step and a store step, which I've included in this commit.
A store to a union field creates a union `Content` containing
the union as a column instead of the field. This means that
we'll correctly find this flow:
```cpp
union U { int x; int y; }
...
U u;
u.x = source();
sink(u.y);
```
since both the store and read step will use a
'TUnionContent(U)'.
I omitted the index column in the example, but
it has a similar rolse as in the 'struct' case.