Fixing wrong example
The definition of the ``ancestorOf`` predicate doesn't really describe an ancestor relation, but its inverse (i.e., something one might call ``descendantOf``). The fix replaces ``parentOf`` with ``childOf``, so that the predicate actually describes the ancestor relation. I've tested the predicate against the tutorial dataset, and it now yields the expected results.