Python: illustrate def-nodes for subscript
Since we do not have index labels, we need to make some
assumptions to recover the index.

Also, this particular implementation might be wrong in a case
such as `foo[3] = bar[4] = "baz"`.
Here `baz` defines the subscript of two different objects
(`foo` and `bar`) and we would pick up both even though
only one  of them might be consistent with the API node.

An more precise implementation is

```codeql
      exists(API::Node subscript |
        // these two lines state that we are defining a subscript
        // `subscript.asSource()` does not exist.
        subscript = aiohttpResponseInstance().getMember("cookies").getASubscript() and
        value = subscript.asSink() and
        // here, we recover the object that is being subscripted
        this = subscript.getAPredecessor(API::Label::subscript()).asSource() and
        // and now we can find the index
        // this part assumes a specific syntax for defining the subscript
        exists(SubscriptNode subscriptNode |
          subscriptNode.getObject() = this.asCfgNode() and
          index.asCfgNode() = subscriptNode.getIndex()
        )
      )
```