Ruby: Add custom barrier guard for cases
`when` clauses in case expressions exert control over their
corresponding `then` branches, but this is not easily noticeable in the
CFG. The individual patterns in the clause have conditional successors,
not the `when` node itself.

Specifically, this means that in an expression such as

    case foo
    when "foo", "bar"
      foo
    end

the pattern `"bar"` technically does not control the read of `foo` on
line 3, because that read is also reachable from the previous pattern
`"foo"`.

Instead of considering the patterns in isolation, we instead consider
the entire `when` clause, provided that all patterns have edges to the
`then` node with the same successor.

In practice this means we copy the implementation of
`DataFlow::BarrierGuard` but use a custom predicate in place of
`ConditionBlock.controls`.