Swift: do not treat bodies within `TrapExpr` specially in CFG
Previously there was code that was treating `BraceStmt` within `TapExpr`
in a special way, skipping the first variable declaration element which
was purposedly visited first in the `TapExpr` itself.

This might have introduced CFG inconsistencies, and here this is
switched to a simpler setup where the tap expression variable
declaration is visited in the body as is normal for `BraceStmt`.