Support quoted operands in access path components
Operands in access path components can now be surrounded by single quotes in order to use reserved characters (commas and brackets) as part of the operand text.

As an escape mechanism, two consecutive single quote characters are converted to an individual single quote. This was chosen mainly since other escaping mechanisms are hard to implement efficiently in CodeQL.